+++ plom/solutions/msgr_utils.py
@@ -1,8 +1,0 @@
-# SPDX-License-Identifier: AGPL-3.0-or-later
-# Copyright (C) 2022 Colin B. Macdonald
-
-from plom.create import with_manager_messenger
-
-# And that would be all, except the message would say "plom-create clear"
-# Do we care?  Maybe we'd rather it said "plom-solution clear"/
-# Obvious solution: copy paste the connect manager here