+++ plom_server/Rectangles/services/__init__.py
@@ -1,0 +1,5 @@
+# SPDX-License-Identifier: AGPL-3.0-or-later
+# Copyright (C) 2024 Andrew Rechnitzer
+
+
+from .rectangle import RectangleExtractor, get_reference_rectangle