+++ CHANGELOG.md
@@ -9,13 +9,25 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 ### Added
+
+### Changed
+
+### Fixed
+
+
+
+## [0.15.5] - 2024-06-12
+
+### Added
 * Users can change their own passwords.
 
 ### Changed
 * Plom now allows sharing pages between questions.
+* Client: parameterized rubrics are no longer considered experimental.
 
 ### Fixed
-
+* For now, you must manually create system rubrics before marking begins; to enforce this the client will crash if the server has no rubrics.
+* Misc fixes and documentation improvements.
 
 
 ## [0.15.4] - 2024-05-21
@@ -1183,7 +1195,8 @@ in most cases.
 This is the first release of Plom, Paperless Open Marking.
 
 
-[Unreleased]: https://gitlab.com/plom/plom/-/compare/v0.15.4...main
+[Unreleased]: https://gitlab.com/plom/plom/-/compare/v0.15.5...main
+[0.15.5]: https://gitlab.com/plom/plom/-/compare/v0.15.4...v0.15.5
 [0.15.4]: https://gitlab.com/plom/plom/-/compare/v0.15.3...v0.15.4
 [0.15.3]: https://gitlab.com/plom/plom/-/compare/v0.15.2...v0.15.3
 [0.15.2]: https://gitlab.com/plom/plom/-/compare/v0.15.1...v0.15.2